Simplify execution stats aggregator
