--- pinot-core/src/main/java/org/apache/pinot/core/query/reduce/ExecutionStatsAggregatorV2.java
@@ -1,234 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.core.query.reduce;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
-import java.util.function.LongConsumer;
-import java.util.stream.Collectors;
-import javax.annotation.Nullable;
-import org.apache.pinot.common.datatable.DataTable;
-import org.apache.pinot.common.metrics.BrokerMeter;
-import org.apache.pinot.common.metrics.BrokerMetrics;
-import org.apache.pinot.common.metrics.BrokerTimer;
-import org.apache.pinot.common.response.broker.BrokerResponseNative;
-import org.apache.pinot.common.response.broker.BrokerResponseStats;
-import org.apache.pinot.common.response.broker.QueryProcessingException;
-import org.apache.pinot.core.transport.ServerRoutingInstance;
-import org.apache.pinot.spi.config.table.TableType;
-import org.apache.pinot.spi.utils.builder.TableNameBuilder;
-
-
-public class ExecutionStatsAggregatorV2 {
-  private final List<QueryProcessingException> _processingExceptions = new ArrayList<>();
-  private final Map<String, Map<String, String>> _operatorStats = new HashMap<>();
-  private final Set<String> _tableNames = new HashSet<>();
-  private final Map<String, String> _traceInfo = new HashMap<>();
-  private final Map<DataTable.MetadataKey, Object> _aggregatedStats = new HashMap<>();
-
-  private final boolean _enableTrace;
-
-  public ExecutionStatsAggregatorV2(boolean enableTrace) {
-    _enableTrace = enableTrace;
-  }
-
-  public void aggregate(ServerRoutingInstance routingInstance, DataTable dataTable) {
-    aggregate(routingInstance, dataTable.getMetadata(), dataTable.getExceptions());
-  }
-
-  public synchronized void aggregate(@Nullable ServerRoutingInstance routingInstance, Map<String, String> metadata,
-      Map<Integer, String> exceptions) {
-
-    // Reduce on trace info.
-    if (_enableTrace && metadata.containsKey(DataTable.MetadataKey.TRACE_INFO.getName())) {
-      _traceInfo.put(routingInstance.getShortName(), metadata.get(DataTable.MetadataKey.TRACE_INFO.getName()));
-    }
-
-
-    for (Map.Entry<String, String> entry : metadata.entrySet()) {
-      String key = entry.getKey();
-      String value = entry.getValue();
-
-      DataTable.MetadataKey metadataKey = DataTable.MetadataKey.getByName(key);
-
-      if (metadataKey == null) continue;
-      if (metadataKey == DataTable.MetadataKey.TRACE_INFO) continue;
-
-      switch (metadataKey.getValueType()) {
-        case INT:
-        case LONG: {
-          _aggregatedStats.put(metadataKey,
-              (Long) _aggregatedStats.getOrDefault(metadataKey, 0L) + Long.parseLong(value));
-          break;
-        }
-        case STRING: {
-          try {
-            boolean boolVal = Boolean.parseBoolean(value);
-            boolean existingBoolVal =
-                Boolean.parseBoolean((String) _aggregatedStats.getOrDefault(metadataKey, "false"));
-            _aggregatedStats.put(metadataKey, String.valueOf(existingBoolVal | boolVal));
-          } catch (Exception e) {
-            List<String> list = (List<String>) _aggregatedStats.getOrDefault(metadataKey, new ArrayList<>());
-            list.add(value);
-            _aggregatedStats.put(metadataKey, list);
-          }
-          break;
-        }
-      }
-    }
-
-    String tableNamesStr = metadata.get(DataTable.MetadataKey.TABLE.getName());
-    if (tableNamesStr != null) {
-      List<String> tableNames = Arrays.stream(tableNamesStr.split("::")).collect(Collectors.toList());
-      _tableNames.addAll(tableNames);
-    }
-
-    // Reduce on exceptions.
-    for (int key : exceptions.keySet()) {
-      _processingExceptions.add(new QueryProcessingException(key, exceptions.get(key)));
-    }
-
-  }
-
-  public void setStats(BrokerResponseNative brokerResponseNative) {
-    setStats(null, brokerResponseNative, null);
-  }
-
-  public void setStats(@Nullable String rawTableName, BrokerResponseNative brokerResponseNative,
-      @Nullable BrokerMetrics brokerMetrics) {
-    // set exception
-    List<QueryProcessingException> processingExceptions = brokerResponseNative.getProcessingExceptions();
-    processingExceptions.addAll(_processingExceptions);
-
-    // add all trace.
-    if (_enableTrace) {
-      brokerResponseNative.getTraceInfo().putAll(_traceInfo);
-    }
-
-    // Set execution statistics.
-    brokerResponseNative.setNumDocsScanned(getLongValue(DataTable.MetadataKey.NUM_DOCS_SCANNED));
-    brokerResponseNative.setNumEntriesScannedInFilter(getLongValue(DataTable.MetadataKey.NUM_ENTRIES_SCANNED_IN_FILTER));
-    brokerResponseNative.setNumEntriesScannedPostFilter(getLongValue(DataTable.MetadataKey.NUM_ENTRIES_SCANNED_POST_FILTER));
-    brokerResponseNative.setNumSegmentsQueried(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_QUERIED));
-    brokerResponseNative.setNumSegmentsProcessed(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_PROCESSED));
-    brokerResponseNative.setNumSegmentsMatched(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_MATCHED));
-    brokerResponseNative.setTotalDocs(getLongValue(DataTable.MetadataKey.TOTAL_DOCS));
-    brokerResponseNative.setNumGroupsLimitReached(Boolean.parseBoolean(
-        (String) _aggregatedStats.getOrDefault(DataTable.MetadataKey.NUM_GROUPS_LIMIT_REACHED, "false")));
-
-    brokerResponseNative.setNumSegmentsPrunedByServer(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_PRUNED_BY_SERVER));
-    brokerResponseNative.setNumSegmentsPrunedInvalid(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_PRUNED_INVALID));
-    brokerResponseNative.setNumSegmentsPrunedByLimit(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_PRUNED_BY_LIMIT));
-    brokerResponseNative.setNumSegmentsPrunedByValue(getLongValue(DataTable.MetadataKey.NUM_SEGMENTS_PRUNED_BY_VALUE));
-    brokerResponseNative.setExplainPlanNumEmptyFilterSegments(getLongValue(DataTable.MetadataKey.EXPLAIN_PLAN_NUM_EMPTY_FILTER_SEGMENTS));
-    brokerResponseNative.setExplainPlanNumMatchAllFilterSegments(getLongValue(DataTable.MetadataKey.EXPLAIN_PLAN_NUM_MATCH_ALL_FILTER_SEGMENTS));
-
-    brokerResponseNative.setNumConsumingSegmentsQueried(getLongValue(DataTable.MetadataKey.NUM_CONSUMING_SEGMENTS_QUERIED));
-    brokerResponseNative.setNumConsumingSegmentsProcessed(getLongValue(DataTable.MetadataKey.NUM_CONSUMING_SEGMENTS_PROCESSED));
-    brokerResponseNative.setNumConsumingSegmentsMatched(getLongValue(DataTable.MetadataKey.NUM_CONSUMING_SEGMENTS_MATCHED));
-
-    if (_aggregatedStats.containsKey(DataTable.MetadataKey.MIN_CONSUMING_FRESHNESS_TIME_MS)) {
-      brokerResponseNative.setMinConsumingFreshnessTimeMs(
-          getLongValue(DataTable.MetadataKey.MIN_CONSUMING_FRESHNESS_TIME_MS));
-    }
-
-    // OFFLINE/REALTIME
-    String tableName = _tableNames.isEmpty() ? rawTableName : _tableNames.iterator().next();
-    TableType tableType = null;
-    if (tableName != null && tableName.isEmpty()) {
-      tableType = TableNameBuilder.getTableTypeFromTableName(tableName);
-    }
-
-    if (tableType == TableType.OFFLINE) {
-      brokerResponseNative.setOfflineThreadCpuTimeNs(getLongValue(DataTable.MetadataKey.THREAD_CPU_TIME_NS));
-      brokerResponseNative.setOfflineSystemActivitiesCpuTimeNs(getLongValue(DataTable.MetadataKey.SYSTEM_ACTIVITIES_CPU_TIME_NS));
-      brokerResponseNative.setOfflineResponseSerializationCpuTimeNs(getLongValue(DataTable.MetadataKey.RESPONSE_SER_CPU_TIME_NS));
-      brokerResponseNative.setOfflineTotalCpuTimeNs(
-          brokerResponseNative.getOfflineThreadCpuTimeNs() + brokerResponseNative.getOfflineSystemActivitiesCpuTimeNs()
-              + brokerResponseNative.getOfflineResponseSerializationCpuTimeNs()
-              + brokerResponseNative.getOfflineThreadCpuTimeNs());
-    }
-
-    if (tableType == TableType.REALTIME) {
-      brokerResponseNative.setRealtimeThreadCpuTimeNs(getLongValue(DataTable.MetadataKey.THREAD_CPU_TIME_NS));
-      brokerResponseNative.setRealtimeSystemActivitiesCpuTimeNs(getLongValue(DataTable.MetadataKey.SYSTEM_ACTIVITIES_CPU_TIME_NS));
-      brokerResponseNative.setRealtimeResponseSerializationCpuTimeNs(getLongValue(DataTable.MetadataKey.RESPONSE_SER_CPU_TIME_NS));
-      brokerResponseNative.setRealtimeTotalCpuTimeNs(
-          brokerResponseNative.getRealtimeThreadCpuTimeNs() + brokerResponseNative.getRealtimeSystemActivitiesCpuTimeNs()
-              + brokerResponseNative.getRealtimeResponseSerializationCpuTimeNs());
-    }
-
-    // Update broker metrics.
-    if (brokerMetrics != null && rawTableName != null) {
-      addBrokerMetrics(rawTableName, brokerMetrics, brokerResponseNative);
-    }
-  }
-
-  private void addBrokerMetrics(String rawTableName, BrokerMetrics brokerMetrics, BrokerResponseNative brokerResponseNative) {
-    brokerMetrics.addMeteredTableValue(rawTableName, BrokerMeter.DOCUMENTS_SCANNED,
-        brokerResponseNative.getNumDocsScanned());
-    brokerMetrics.addMeteredTableValue(rawTableName, BrokerMeter.ENTRIES_SCANNED_IN_FILTER,
-        brokerResponseNative.getNumEntriesScannedInFilter());
-    brokerMetrics.addMeteredTableValue(rawTableName, BrokerMeter.ENTRIES_SCANNED_POST_FILTER,
-        brokerResponseNative.getNumEntriesScannedPostFilter());
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.OFFLINE_THREAD_CPU_TIME_NS,
-        brokerResponseNative.getOfflineThreadCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.REALTIME_THREAD_CPU_TIME_NS,
-        brokerResponseNative.getRealtimeThreadCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.OFFLINE_SYSTEM_ACTIVITIES_CPU_TIME_NS,
-        brokerResponseNative.getOfflineSystemActivitiesCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.REALTIME_SYSTEM_ACTIVITIES_CPU_TIME_NS,
-        brokerResponseNative.getRealtimeSystemActivitiesCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.OFFLINE_RESPONSE_SER_CPU_TIME_NS,
-        brokerResponseNative.getOfflineResponseSerializationCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.REALTIME_RESPONSE_SER_CPU_TIME_NS,
-        brokerResponseNative.getRealtimeResponseSerializationCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.OFFLINE_TOTAL_CPU_TIME_NS,
-        brokerResponseNative.getOfflineTotalCpuTimeNs(), TimeUnit.NANOSECONDS);
-    brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.REALTIME_TOTAL_CPU_TIME_NS,
-        brokerResponseNative.getRealtimeTotalCpuTimeNs(), TimeUnit.NANOSECONDS);
-
-    if (_aggregatedStats.containsKey(DataTable.MetadataKey.MIN_CONSUMING_FRESHNESS_TIME_MS)) {
-      brokerMetrics.addTimedTableValue(rawTableName, BrokerTimer.FRESHNESS_LAG_MS,
-          System.currentTimeMillis() - brokerResponseNative.getMinConsumingFreshnessTimeMs(), TimeUnit.MILLISECONDS);
-    }
-  }
-
-  public void setStageLevelStats(@Nullable String rawTableName, BrokerResponseStats brokerResponseStats,
-      @Nullable BrokerMetrics brokerMetrics) {
-    setStats(rawTableName, brokerResponseStats, brokerMetrics);
-
-    brokerResponseStats.setNumBlocks(getLongValue(DataTable.MetadataKey.NUM_BLOCKS));
-    brokerResponseStats.setNumRows(getLongValue(DataTable.MetadataKey.NUM_ROWS));
-    brokerResponseStats.setStageExecutionTimeMs(getLongValue(DataTable.MetadataKey.OPERATOR_EXECUTION_TIME_MS));
-    brokerResponseStats.setOperatorStats(_operatorStats);
-    brokerResponseStats.setTableNames(new ArrayList<>(_tableNames));
-  }
-
-  private long getLongValue(DataTable.MetadataKey metadataKey) {
-    return (Long) _aggregatedStats.getOrDefault(metadataKey, 0L);
-  }
-
-}
